saturncc: signed int types eliminate LOAD node, r1 for constants
Change FUN_06004378's C source from `unsigned int c` / `unsigned int u`
to `int c` / `int u`. This eliminates the unsigned→signed LOAD node
that was consuming a scratch register, shifting all subsequent
allocations from r1 to r2.

With the LOAD gone, the first available scratch (r1) goes directly
to comparison constants. The second constant in each range check
(0x39, 0x5A) now produces byte-identical output to the Hitachi
reference: `e1 39` / `3e 17` match perfectly. The first constant
still uses r1 where the reference uses r3 — an allocator difference.

The BANDI4(reg,bmask) rule already existed for `extu.b` with signed
operands, so `int u = c & 0xff` compiles to `extu.b r4,r14` directly.

FUN_06004378 range check 1: 17/24 bytes match (up from 16/24).
7 of 12 instruction pairs are now fully byte-identical.

+++ saturn/experiments/daytona_byte_match/FUN_06004378.c
@@ -26,11 +26,11 @@
  * `mov.w @(disp,PC)`. The actual values from the binary are 0x00D0
  * and 0x00C9 — known short constants, not runtime dereferences. */
 
-int FUN_06004378(unsigned int c) {
-    unsigned int u = c & 0xff;
-    if ((int)u >= 0x30 && (int)u <= 0x39)
+int FUN_06004378(int c) {
+    int u = c & 0xff;
+    if (u >= 0x30 && u <= 0x39)
         return 0x00D0 + c;
-    if ((int)u >= 0x41 && (int)u <= 0x5a)
+    if (u >= 0x41 && u <= 0x5a)
         return 0x00C9 + c;
     if (u == 0x22) return 0x25;
     if (u == 0x27) return 0x24;
